7.2. Refactor actions file structure
