+++ src/store/reducers/data/data.js
@@ -1,4 +1,4 @@
-import {ActionType} from '../../action.js';
+import {ActionType} from '../../actions/actions.js';
 import {extend} from '../../../utils/common.js';
